clinker-core: split aggregation.rs into focused submodules
The 4.7k-line aggregation.rs mixed six unrelated concerns behind no
module boundary, so a change to spill serialization forced a recompile
of streaming logic and contributors had to read through spill writers
to understand hash-vs-streaming behavior.

Convert it to an aggregation/ directory:

  mod.rs   — re-export hub plus the residual evaluator, the
             AggregateStream dispatch wrapper, the AccumulatorOp
             ingestion modes, the StreamingAggregator, the sidecar-merge
             helpers, and the plan-time streaming-eligibility qualifier.
  error.rs — the AggregateStrategy plan enum and the AggregateEvalError /
             HashAggError taxonomy plus its PipelineError mapping (the
             lowest layer every other submodule depends on).
  hash.rs  — the HashAggregator hash table, the AccumulatorFactory
             prototype clone factory, the per-group memory estimator,
             the AggregateConsumer arbitrator hook, and the shared
             group-finalize helpers the streaming path reuses.
  spill.rs — the binary postcard+LZ4 spill writer/reader plus the
             external-merge entry and on-disk payload types.

Pure relocation: no symbol renamed, no behavior changed, and the
clinker_core::aggregation::* public surface is preserved verbatim via
re-exports. File-private helpers that now straddle the submodule
boundary are widened only to pub(super); no test was altered — the
spill-trigger test module that white-boxes HashAggregator internals
moves into hash.rs so those fields stay private.

Closes #127
